Some validation and Authorisation changes

+++ Tournament/Tournament/Controllers/PlayerController.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
+using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -22,51 +24,86 @@ namespace TournamentApi.Controllers
         {
             _context = context;
         }
-        [Authorize(Roles = "User,Moderator,Admin")]
         // GET: api/Player
+        [Authorize(Roles = "Moderator,Admin")]
         [HttpGet]
         public async Task<ActionResult<IEnumerable<Player>>> GetPlayers()
         {
             return await _context.PlayerItems.ToListAsync();
         }
 
-        [Authorize(Roles = "User,Moderator,Admin")]
         // GET: api/Player/{id}
+        [Authorize(Roles = "User,Moderator,Admin")]
         [HttpGet("{id}")]
         public async Task<ActionResult<Player>> GetPlayers(long id)
         {
-            var playerItem = await _context.PlayerItems.FindAsync(id);
-
-            if (playerItem == null)
+            var identity = HttpContext.User.Claims.ToList();
+            long userId;
+            if (identity[0].Value != "null")
+                userId = long.Parse(identity[0].Value);
+            else
+                userId = -99;
+            var role = identity[1].Value;
+            var clientId = long.Parse(identity[2].Value);
+            if ((role == "User" && id == clientId))
             {
-                return NotFound();
+                var playerItem = await _context.PlayerItems.Include(p => p.user).Where(p => p.Id == id).FirstOrDefaultAsync();
+                playerItem.user.Password = null;
+                if (playerItem == null)
+                {
+                    return NotFound();
+                }
+                return playerItem;
+            }
+            else
+            {
+                return Unauthorized();
             }
 
-            return playerItem;
         }
-        [Authorize(Roles = "User,Admin")]
         // POST: api/Player
+        [Authorize(Roles = "User")]
         [HttpPost]
         public async Task<ActionResult<Player>> PostPlayer(Player item)
         {
-            _context.PlayerItems.Add(item);
-            await _context.SaveChangesAsync();
+            var identity = HttpContext.User.Claims.ToList();
+            long clientId = long.Parse(identity[2].Value);
+            if (_context.PlayerItems.Include(u => u.user).Where(c => c.user.Id == clientId).FirstOrDefault() != null)
+            {
+                return BadRequest("You already have player profile");
+            }
+            else if(_context.PlayerItems.Include(u => u.user).Where(c => c.InGameName == item.InGameName).FirstOrDefault() != null)
+                {
+                return BadRequest("Player with this ingame name already exists");
+            }else
+            {
+                item.user = _context.Users.Where(u => u.Id == clientId).FirstOrDefault();
+                _context.PlayerItems.Add(item);
+                await _context.SaveChangesAsync();
 
-            return CreatedAtAction(nameof(GetPlayers), new { id = item.Id }, item);
+                return Ok("User created");
+            }
         }
-        [Authorize(Roles = "User,Admin")]
         // PUT: api/Player/{id}
+        [Authorize(Roles = "User")]
         [HttpPut("{id}")]
         public IActionResult PutPlayer(long id, Player item)
         {
+            
             var existingPlayer = _context.PlayerItems.Where(s => s.Id == id).FirstOrDefault<Player>();
+            var identity = HttpContext.User.Claims.ToList();
+            long userId;
+            if (identity[0].Value != "null")
+                userId = long.Parse(identity[0].Value);
+            else
+                userId = -99;
+            var role = identity[1].Value;
 
-            if (existingPlayer != null)
+            if (existingPlayer != null && userId == id)
             {
                 existingPlayer.InGameName = item.InGameName;
                 existingPlayer.Region = item.Region;
                 existingPlayer.Username = item.Username;
-
                 _context.SaveChanges();
                 return NoContent();
             }
@@ -76,11 +113,18 @@ namespace TournamentApi.Controllers
             }
             
         }
-        [Authorize(Roles = "User,Moderator,Admin")]
         // DELETE: api/Player/{id}
+        [Authorize(Roles = "User,Moderator,Admin")]
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeletePlayer(long id)
         {
+            var identity = HttpContext.User.Claims.ToList();
+            long userId;
+            if (identity[0].Value != "null")
+                userId = long.Parse(identity[0].Value);
+            else
+                userId = -99;
+            var role = identity[1].Value;
             var playerItem = await _context.PlayerItems.FindAsync(id);
 
             if (playerItem == null)
@@ -88,10 +132,15 @@ namespace TournamentApi.Controllers
                 return NotFound();
             }
 
-            _context.PlayerItems.Remove(playerItem);
-            await _context.SaveChangesAsync();
+            if (id == userId || role == "Admin")
+            {
+                _context.PlayerItems.Remove(playerItem);
+                await _context.SaveChangesAsync();
 
-            return NoContent();
+                return NoContent();
+            }
+            else
+                return BadRequest();
         }
     }
 }
